requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
chromadb==0.4.18
sentence-transformers==2.2.2
pyairtable==2.2.0
python-multipart==0.0.6